Added the skeleton code of the query of finding unsafe objects.
The first step of our sandboxing unsafe Rust project is to identify
all unsafe memory objects. I originally planned to do this in an MIR
pass, but later found that may not be the best implementation choice
because of the how rustc is organized. We should collect interested data
after the compiler has run all its basic transformation passes, and we
should find a way to assembl the data of each function in a crate and
write them all to a file for the later summary-based inter-procedural
analysis. Intuitively we can create a collection data structure for the
current crate to hold all the data; however, Rust's MIR only support
function passes and there is no such an mir::crate struct.
rustc uses a query-based compiling strategy (https://rustc-dev-guide.rust-lang.org/query.html)
which serves the dual purposes of building and querying compliation
results (it caches data queried before). Therefore we should use the
query system to store our interested data. However, we cannot call a
query function at the end of the optimization pass pipeline (https://github.com/jzhou76/surust/blob/2a1b106541187fc0f2b70c68fa763812c97fba70/compiler/rustc_mir_transform/src/lib.rs#L510):
We need a reference to the mir::Body of a fn to iterate over and compute
the data we need, but a query function cannot accept an argument of a
reference to Body (usually accepting a rustc_span::def_id::DefId).
Inside a query function we can get the Body of a function by calling
optimized_mir() with a DefId, but this does not work for us because
it woulc cause a cycle in the call graph (optimized_mir() -> our_pass()
-> query_function() -> optimized_mir()).

The current solution is to directly call the query function after all
optimization passes finish. The callsite is in function
collect_neighors() at rustc_monomorphize/src/collector.rs.
It works, but I'm not certain this is the best place to do it.