Moved unsafe_obj_mir() call from collect_neighbours to collect_items_rec
There should be no differences in the generated code because there are
no code transformations between the old callsite and new callsite.
However, semantically it looks weird to put the unsafe object collection
procedure inside a function that "collect neighbours".

However, the new callsite still may not be the best place. Probably
we should call the sandboxing related procedures at the end of
collect_items_rec().