[find_unsafe_obj] Refactored code to use a normal fn instead of query.
At a high-level, the first step of the sandboxing work is to collect
unsafe object related information intra-procedurally. We will collect
not only the allocation site of each unsafe object, but also a lot of
other data such as which function calls are made. We should gather all
the desired information into one struct and return it for later
inter-procedural analysis. However, under the query system, this cannot
be done conveniently: there would be a dependency cycle between the
rustc_middle crate and the rustc_mir_transform crate if we define the
aforementioned struct in the transform crate; and it would be unclean to
define such a struct in the rustc_middle as it is intended to define
fundamental data structures and functions. Also, we only need to call
the procedure of collecting unsafe object information once, there is no
need to use the query system and a normal function call to it should
work.

TODO: Currently we only collect unsafe objects from each function (WIP),
but as mentioned above, ultimately we need collect much more information
for later inter-procedural analysis.