[WIP] Commit WIP changes before mergeing rust-1.60.0.
The dependent libraries for the compiler are constantly under updating.
For example, after I `cargo update` today, the compiler failed to
compile because of the follow error:

error[E0412]: cannot find type `c_char` in module `core::ffi`
  --> /Users/jie/.cargo/registry/src/github.com-1ecc6299db9ec823/compiler_builtins-0.1.72/src/mem/mod.rs:74:58
   |
74 |     pub unsafe extern "C" fn strlen(s: *const core::ffi::c_char) -> usize {
   |                                                          ^^^^^^ not found in `core::ffi`

the reason is that the newest core library defines core::ffi::c_char
while the current library (1.58.1) does not. I plan to solve this
(and future similar) problem by adding the original rust remote to
this repo, and merge the new changes from the vanillia repo to our
surust repo periodically to keep it up to date.