[heap_alloc] Wrote code to find heal allocation calls.
Currently the way of finding heap allocation calls is pretty naive:
We simply check if the name of the called the function is in a preset
list of heap allocation calls such as "new" and "with_capacity". There
are two potential problems here:

1. The list may be incomplete.
2. We only check the name of the method without checking where this
method is from (e.g., only "new" from "Box::<i32>::new()"). This poses
the possibiliy of introducing false positives.