[handle_unsafe_op_core] Fixed a bug in computing statement index.
The procedure examines a BB backward, starting from the last unsafe
operation or the terminator of the BB. When it first examined a
terminator, it decremented the index of the terminator and useed the
updated index as the starting point for iterating over the statements.
However, if the BB only has one terminator and no statemtns, the index
would underflow to usize::MAX. This commit fixed this bug by checking
the number of statements before iterating over them.