[handle_unsafe_op_core] Check if an unsafe Place is from fn arguments.
After examing the entry BB, if there are unsafe Place unhandled in
the unsafe Place set, we should check if they are from the function's
arguments.

Also fixed a bug in extracting Place in return value of a function.
Previous code only counts non-empty Tuple. What we want is to exclude
empty Tuple and keep others.