[find_unsafe_op_core] Minor optimization for compile time.
When recursively calling the find_unsafe_core function on a BB's
predecessors, if there are multiple predecessors, we need pass a clone
of the current unsafe Place set for each predecessor.  For example,
the unsafe Place _2 may be defined in both of the two predecessors by
"_2 = Vec::new()". Simply passing the reference instead of a clone would
miss the definition in one branch.  However, if there is only one
predecessor, we can just pass the mutable reference of the current
unsafe Place set to it.