Moved the call to find_unsafe_obj to after collecting mono items.
Previously in the monomorphize module, the find_unsafe_obj module is
called during collecting mono items. This commit moved the call to
after collecting all the mono items. This should not improve the compile
time as the previous already ignored visited items during collecting;
however, the current code is cleaner. After the call, the compiler
starts to partition MIR for later LLVM code gen; so this call should be
at the latest point.