[summarize_fn] Wrote the skeleton code of summarizing each function.
Wrote the skeleton code of summarizing each function for later
whole-program analysis. Since function calls connect different
functions, we defined a struct summarize_fn::Summary that contains the
information about the callees used in a function. The struct
summarize_fn::Callee has the allocation sites for all the arguments
used by the callees.

This commit also temporarily turns off the analysis of finding
allocation sites of each Place used in unsafe blocks. It should be
cleaner to do that analysis after the finally summary is generated.

TODO: Collect the allocation sites of the arguments of each callee.