[summarize_fn] Fixed a bug in finding definition sites of call args.
There is a corner case that breaks the previous implemention: A variadic
callee is called more than once with different number of arguments AND
in the initialization phase the fewer-arg call is processed before the
more-arg call. In this case, Callee would only reserve the fewer number
of spots for its arguments, and thus later when processing a callee with
more arguments, there would be out-of-bounds accesses. This commit fixed
this problem by dynamically allocating one more spot for the arg vector.
See the code for more details.

Also changed "(A)alloc" to "(D)def" in many places because we are
looking for allocation/declaration sites and "definition" is more
accurate than only "allocation".