Fixed a bug in iterating over all MonoItem.
A generic function has only one DefId, even if it is instantiated to
multiple MonoItem. When iterating over all the MonoItem, we recorded
which function has been processed and skip the processed. We should
use DefId as the unique ID instead of DefId.index (function's local ID).

Also moved the invocation of the summarize_fn in collector.rs from
inside collect_crate_mono_items() to a separate function.