[summarize_fn] Write summaries of dependency crates to temporary files.
Write the fn summaries of all the dependency crates for an app to
temporary files. Those files will later be read by the main crate for
inter-procedural analysis.

Currently we assume that all crates used by a Rust app, including the
app iteself, will be compiled by a rustc that is invoked by the same
cargo, as long as it is built as "cargo build ...".  Based on this
assumption, we use a temporal directory "/tmp/rust-sandbox-cargo_pid"
to store all the summary files. However, for some programs such as
RustPython, we observed that there might be some extra directories
generated with files named probe1/2/3/... They do not look like crates
for the target app (some of them are empty), but I do not know what
they are exactly.