[summarize_fn] Categorize the call in DefSite.
Originally DefSite only has one LocBB(u32) to represent the location of
a call. However, this is not enough because when doing WPA, we need to
know the exact type of the call. For example, when we see a Place _10
is used in unsafe code, and find that _10 is defined by some function
call, i.e., _10 = func(...), we need to figure out what func is.
If func is a heap allocation function such as Vec::new(), then we have
found an unsafe heap allocation site. If func is a native library
function, no further actions will be executed because we do not analyze
native library functions. Otherwise, func is a normal function and the
analysis should go to query what contribute(s) to the return value of
func. Note that there is no need to analyze the arguments of func in
WPA because unsafe_def has done it.

Also made many other revisions, including renaming functions and
improving comments.