[summarize_fn] Revised code so later WPA will be more precise.
Improved both the unsafe_def and calls modules so that later WPA can
be more precise in indentifying unsafe allocations. Specifically,
this commit made two major changes.

First, the old unsafe_def::find_unsafe_def_core conservatively assumes
that every argument of a normal function call may contribute to the
return value, which is an overesimate. However, we do not know which
exact argument(s) contribute to the return value until we have run
summarize_fn::calls on the callee. This commit updated
find_unsafe_def_core to not search for the def sites of each argument
and only record the call (technically, the location of the call).
We will not lose necessary information because summarize_fn::calls
computes the def sites of each argument of each call. We wait for the
WPA to query the summary from the calls module to only search for the
def sites of argument(s) that contribute to a target return value.

Second, the old Callee.def_sites does not distinguish each call to the
same callee, and thus the def sites for the argument of a callee is
a superset of the def sites for a specific call. This is useful when
we search for the def sites of a function argument that contributes
to an unsafe Place. For example, _2 is used in unsafe code of function
foo, and _2 is the second argument of foo. We need to find all callers
of foo, and in each caller, find all the def sites of the second
argument of all calls to foo. In this case, there is no need to
distinguish the call sites. However, in the scenario described in the
paragraph above, say function foo has "_10 = bar(_8, ...);" in bb9, it
will be more precise to only find the def sites for _8 in the context
of the call in bb9, rather than to find the def sites of the first
argument of all foo's calls to bar. This commit updated Callee.def_sites
from Vec::<FxHasSet<DefSite>> to HashMap<BB, Vec::<FxHasSet<DefSite>>>
to support this.

Also updated Summary.ret_defs to distinguish DefSite for calls and args.

Also made some refactoring to improve code quality.