Fixed a concurrency problem in creating the directory for fn summaries.
Whichever compilation process that finishes its local analysis first can
create the directory for all the temporary summary files for all
dependent crates. Previously summarize_fn::mod checks if this directory
exists before creating one, and panics when the creation fails.
There is a bug: multiple processes could pass the Path::exists() check
and try to create the directory at the same time, which would panic
the compilation with the "File exists" error. It is actually harmless
to simply call fs::create_dir() without caring about the return code.
If none of the processes creates the directory successfully, the
following code of writing results to a file in the directory will panic.