Fixed a file sychronization problem.
The previous code has a major misunderstanding: It assumes that each
crate is compiled only once, which is incorrect. One crate may be
compiled multiple times and each one has a unique StableCrateId.
We can get the list of compiled dependency crates by examining
`cargo tree`. Specifically, cargo will compile the crates in the
normal dep. tree once, and also the crates in the build-dependencies.
This commit fixed this problem as follows. In the main compiling
process, it invokes a python script which internally queries
`cargo tree` to count the number of compiled dependency crates,
and it counts the current number of generated summary files. It does
a busy-waiting to check if all the summary files are ready.

Note that the python script is in the SURust-Misc repository, and the
Rust code hardcodes the path of the python script.  Ideally we should
directly write the logic of the script inside the main compiling
process. Currently the script is only for fast development.