Move the sanboxing project to Rust-1.63.0.
This commit *manually* moved all the changes we did on Rust-1.58.1 to
Rust-1.63.0. This is mainly because we need to use some functionality
(see next commit for more details) which is only available in newer
compilers. Also the doc is always up to date with the newest compiler;
so sometimes we saw inconsistencies between the source code of 1.58.1
and the doc of the newest compiler.

Ideally, we should do `git merge` or `git rebase` for this upgrade work.
However, I tried both and got tons of conflicts that did not seem to
solvable easily (e.g., many conflicts are in files that are not modified
by me). I manually applied all the changes to Rust-1.63.0, which is
the newest compiler to date. I also made some changes due to API
updates or new compiling rules enforced by the newer compiler. Notably:

- The `crate` visibility modifier is not supported anymore (at least
  in the compiler source code). The error message is
  "`crate` visibility modifier is experimental".
- We have to manually put a lifetime parameter for functions that use
  the lifetime parameter for its argument. For example, utils:ignore_fn.
  We also need to put a lifetime parameter for struct implementation,
  e.g., wpa::CallGraph. Missing lifetime parameters causes:
  "error[E0261]: use of undeclared lifetime name `'xxx`".
  I don't understand why this seemingly redundant declaration of
  lifetime parameter is needed now while previously such lifetime
  information can be inferred automatically.

Previous commit history: https://github.com/URSec/SURust/commits/surust-1.58.1

We are still in the middle of implementing the whole-program analysis
that finds unsafe allocations.