Resolve callee that is a trait function.
When calling a trait fn, the DefId returned by callee.literal.ty.kind()
is always the fn declaration or default impl in the trait. We need to
resolve the callee as precisely as possible, otherwise the later
whole-program analysis may use wrong callees. Essentially, we rely on
Instance::resolve() to resolve the concrete callee for us.  However,
sometimes it is impossible to statically get the exact callee
when the call is through a trait object. For such cases, we need to
conservatively assume all the implementors of the trait are possible,
and so we may get a set of candidate callees for one call.

The callee resolving function uses TyCtxt::impl_item_implementor_ids()
to help resolve callees that are not resolved by Instance::resolve().
This API is not available in Rust-1.58.1 and it is the main reason
why we migrated from 1.58.1 to 1.63.0.

Note that this commit does not handle all types of InstanceDef. It is
fine for the monent as none of our test programs trigger the panic
that is used to "guard" the unhandled cases.

See the code for more details.