[calls] Handle foreign callees.
For foreign callees, usually those declared in `extern "C"`, the WPA
will ignore them because rustc does not compile them and thus our local
analysis will not process them.  The previous local analysis of
collecting information about callees did not take this into
consideration and thus when the later WPA tries to find the Summary of
a foreign callee, it will fail. This commit fixed this problem by
recording all the foreign functions called by a function. Another
implementation option is to directly ignore adding such a callee to
the Summary of the caller. However, we chose to add it anyway for
the sake of the completeness of the call graph.