[calls] Record callees that cannot be resolved statically.
Calls via trait objects may not be precisely resolved at compile time,
so our analysis may find multiple callee candidates for a call. We
should assume that all of them are possible call targets (0299b42).
However, some of the candidate calls found my our callee-resolving
procedure (calls::resolve_callee()) may not appear as an MonoItem
in our local analysis.  For example, in crate futures_task-0.3.21,
trait fn UnsafeFutureObj::into_raw() is implemented 10 times, and our
resolving procedure finds all the 10 impl. However, summarize::summarize()
only sees 4 of them. We need to record which callees are
dynamically dispatched so that in later WPA when it cannot find a
Summary for a callee, it can check whether it is caused by this.