[WPA] Handle DefSite::OtherCall in find_unsafe_alloc().
The def site of the return value of a DefSite::OtherCall comes from two
types of places. The first is the arguments of the callee, which come
from the def sites in the body of the caller; and the second is the
non-argument def sites in the body of the callee. This commit finished
finding def sites for DefSite::OtherCall. And it adds comments to
find_unsafe_alloc().