[sandbox_cmd_option] Added a compile flag for the unsafe sandboxing.
Added a compile option for turning on the sandboxing. We added it as a
codegen option. Codegen might not be the best place to add our flag
because codegen options seem to be more relavent with LLVM codegen,
while our analysis and transformation happens on MIR. Another choice
is to add it as an unstable debugging option ("-Z").  However, this
our sandboxing passes are not for debugging.