[WPA] Find unsafe function arguments and non-heap-alloc calls.
Find unsafe fn arguments and non-heap-alloc calls that return unsafe
value.  Combined with find_unsafe_alloc, this prepares unsafe sources
for later local analysis to find unsafe memory accesses.
Specifically, in addition to the unsafe heap allocation sites that were
found before by find_unsafe_alloc, it finds two more unsafe sources:
unsafe function arguments and unsafe non-heap-alloc calls.

Similar to find_unsafe_alloc, find_unsafe_arg_call also uses a
worklist-based algorithm. It starts with all the unsafe heap allocation
sites found before.  It examines two situations. First, if an unsafe
souce is used as an argument to a non-heap-alloc call, the argument of
the callee is also an unsafe resouce and the algorithm puts the argument
to the worklist.  Second, if an unsafe source contributes to the return
value of the function that contains it, then for all the callers of this
function, the calls to it are also unsafe sources.