[unsafe_access] Impl. local analysis to find unsafe memory accesses.
Implemented the module of finding unsafe memory accesses using the
unsafe sources (unsafe heap allocations, arguments, and non-heap-alloc
calls) from the previous WPA. The procedure is very simple. First,
it finds all unsafe Place in a function. A Place on the LHS of a
StatementKind::Assign is unsafe if any Place used to compute this Place
is unsafe. Second, it just check each Statement/Terminator to see if
there are dereferences to unsafe Place.